refactor(ui): move case Search + Status into the right pane, widen rail
The Search and Status filters only affect the test-case list in the right
pane, but lived in the left rail — filtering-at-a-distance. Move them into
the case list's sticky header, next to what they filter. The left rail now
holds only what scopes the whole report (dialect, language, in-scope list)
and is widened 248→280px so the implementation list is less cramped.
